backend/requirements.txt
# --- Core Web Server ---
fastapi
uvicorn[standard]
sse-starlette

# --- AI & Vector DB ---
faiss-cpu
google-generativeai
langchain
langchain-google-genai
langchain-community

# --- GitHub API ---
PyGithub
requests # <--- ADD THIS LINE

# --- Utilities ---
python-dotenv

# --- Local Embeddings (Replaces Google for FAISS) ---
sentence-transformers
langchain-huggingface
# ... all your other libraries ...
cryptography
GitPython